+++ CHANGELOG.md
@@ -45,7 +45,7 @@ See [Inline mode wiki](https://github.com/metafates/mangal/wiki/Inline-mode) for
 </details>
 
 - Improved TUI experience
-- Search completions in TUI
+- Search completions in TUI. `mangal config info -k search.show_query_suggestions`
 - Anilist caching significantly improved. Now, it will cache all search results (for 2 days)
 - Add caching for custom (lua) sources
 - Include different cover sizes and color for json output #116